# Conflicts:
#	src/backend/auth/biz-auth/src/main/kotlin/com/tencent/bkrepo/auth/dao/repository/ProxyRepository.kt
#	src/backend/npm-registry/biz-npm-registry/src/main/kotlin/com/tencent/bkrepo/npm/service/NpmPackageService.kt
#	src/backend/opdata/biz-opdata/src/main/kotlin/com/tencent/bkrepo/opdata/model/GcInfoModel.kt

+++ src/backend/s3/biz-s3/src/main/kotlin/com/tencent/bkrepo/s3/artifact/response/S3ArtifactResourceWriter.kt
@@ -72,7 +72,7 @@ class S3ArtifactResourceWriter (
         val totalSize = resource.getTotalSize().toString()
 
         prepareResponseHeaders(response, totalSize.toLong(), node, status, contentType, characterEncoding, range)
-        response.bufferSize = getBufferSize(range.length.toInt())
+        response.bufferSize = getBufferSize(range.length)
         return writeRangeStream(resource, request, response)
     }
